backend/requirements.txt
alembic==1.15.2
atomicwrites==1.4.1
attrs==25.3.0
click==8.1.8
colorama==0.4.6
Flask==2.2.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==3.2.1
gunicorn==20.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
py==1.11.0
PyJWT==2.10.1
pytest==7.1.2
SQLAlchemy==1.4.46
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==2.2.3
WTForms==3.2.1
